backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pillow==10.2.0
numpy<2
--extra-index-url https://download.pytorch.org/whl/cu121
torch==2.1.2
torchvision==0.16.2
transformers==4.37.0
accelerate==0.26.1
diffusers==0.30.0
safetensors==0.4.1
python-dotenv==1.0.0
openai==1.12.0
requests==2.31.0
insightface==0.7.3
onnxruntime-gpu==1.16.3
huggingface_hub[hf_xet]==0.13.4